Don't recognize elements with textual hrules in AoA; make AoH compatible.
The .from_aoa method used to look inside inner array elements to see if they
looked like hlines, but use in org-mode blocks always just translated hlines to
nils, so no nils are the only way to represent hlines in an array of arrays.

Also, though .from_aoh recognized nils as group boundaries, it did not have a
keyword argument hrules: to turn recognition of hlines on and off. Made it
compatible with .from_aoa, so nils are not allowed by default, only if the
hlines argument is set true.  Needed to expose the hlines keyword argument up
the call chain, i.e., in Table@from_aoh and FatTable.from_aoh.