Make parse_str_fmt -> parse_string_fmt to make it consistent with other types.
Had to rename former parse_string_fmt -> parse_fmt_string to free up the
name.  That method is the top-level parsing of a format string, not the
extraction of String-specific formatting as with the other parse_<type>_fmt
methods are.  And since we call those methods with #send, building the name
out of the column class types, this needed to be made consistent.