Refactor format_numeric method.  Not handling commas correctly.
Symptom: whenever we ask for commas in formatting, we lost the rounding and
displaying to post_digits digits.  This was caused by a change to the
fat_table String#commas method, which would convert the string into a numeric,
then back to a string, losing the post-decimal formatting.

My prior fix was wrong-headed, so will be reverted.