Fix type-dropping problem when operations applied.
In byr, I noticed that a column for stock symbols, which I typed to a string
was still complaining when the AT&T symbol 'T' was added after a select.  The
problem was that, even though the original table has the symbol column
properly typed, the table resulting from the select did not propagate the type
information to the output table.

This fix is designed to address this.  The problem largely stemmed from the
Table#empty_dup method's failure to pass type information along to the dupped
table.  Along the way, we also make sure that the "dup" propagates tolerance
of columns and any instance variables the original table might have had.